Store positions/velocity/height as float32 (schema v33)
The MVD wire carries floating-point coordinates (eighth-unit fixed point,
or true floats under the float-coords extension) that mvd-reader already
decodes as float32, but the analytics pipeline truncated each player
origin to int32 in PositionTrack — losing up to ~1 unit per axis and,
worse, deriving the v32 velocity from those rounded positions.

PositionTrack X/Y/Z, VX/VY/VZ and H are now float32, as are AirgibEvent
heights and the view-layer Position3D/Velocity3D/Hgt and bucket/columnar
position/velocity/height outputs. Velocity is recomputed by central
difference over the full-resolution positions, so the old +/-1-unit
quantization noise is gone.

Values are kept at native float32 resolution in memory; the new
result.Coord / result.Coords types round only the JSON text to 3
decimals (lossless for eighth-unit positions, shedding the float
division/epsilon tail on derived velocity and height). The
PositionTrack.H NoFloor sentinel changes from math.MinInt32 to -1e9
(-1000000000), which float32 represents exactly so it round-trips through
JSON (math.MinInt32 serialized as a float32 became -2147483600).

Bump CurrentSchemaVersion to 33 and regenerate the golden corpus. Time
axes stay int32 ms; view angles stay int16 raw angle16. Docs updated
(RESULT_SCHEMA.md, API.md, READMEs).

+++ mvd-analytics/analyzer/locgraph.go
@@ -177,8 +177,8 @@ func BuildLocGraph(result *Result) *LocGraphResult {
 				continue
 			}
 
-			x := float32(pt.X[i])
-			y := float32(pt.Y[i])
+			x := pt.X[i]
+			y := pt.Y[i]
 
 			// Node-time: residence in this loc grows by the gap to the
 			// next sample (clamped by locgraphSampleDt to avoid death
